refactor: move projectCard to component folder
