+++ src/components/ProjectCard/ProgressCircle/ProgressCircle.module.scss
