+++ src/components/ProjectCard/ProgressCircle/ProgressCircle.tsx
@@ -1,7 +1,7 @@
 import cl from 'classnames';
 import { FC, useEffect, useState } from 'react';
 import { CircularProgressbarWithChildren, buildStyles } from 'react-circular-progressbar';
-import { Text } from '../../../../components/ui/Text/Text';
+import { Text } from '../../ui/Text/Text';
 import s from './ProgressCircle.module.scss';
 
 interface ProgressCircleProps {
