+++ src/components/ProjectCard/ProjectCard.tsx
@@ -1,11 +1,11 @@
 import cl from 'classnames';
 import { FC } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Text } from '../../../../components/ui/Text/Text';
-import { ROUTE_PROFILE_PROJECTS } from '../../../../constants/routesConstants';
-import { TProjectShortInfo } from '../../../../types/data';
-import { convertDateToShortFormat } from '../../../../utils/dateConvertHelpers';
-import { ProgressCircle } from '../ProgressCircle/ProgressCircle';
+import { Text } from '../ui/Text/Text';
+import { ROUTE_PROFILE_PROJECTS } from '../../constants/routesConstants';
+import { TProjectShortInfo } from '../../types/data';
+import { convertDateToShortFormat } from '../../utils/dateConvertHelpers';
+import { ProgressCircle } from './ProgressCircle/ProgressCircle';
 import s from './ProjectCard.module.scss';
 
 interface ProjectCardProps {
@@ -36,7 +36,7 @@ export const ProjectCard: FC<ProjectCardProps> = ({ className = '', project }) =
   };
 
   const handleOnCardTitleTap = () => {
-    if (window.innerWidth < 768) {
+    if (window.innerWidth <= 768) {
       navigate(`${ROUTE_PROFILE_PROJECTS}/${project.id}`);
     }
   };
